kernel-shark: Always use 64 bit variables for timestamps.
All time stamps of the trace records are coded with 64 bits, however
on some systems the size_t type can be 32 bits.

Link: http://lore.kernel.org/linux-trace-devel/20190614135045.17223-3-ykaradzhov@vmware.com
Bugzilla: https://bugzilla.kernel.org/show_bug.cgi?id=203869

Reported-by: Alan Mikhak <alanmikhak@gmail.com>

+++ kernel-shark/src/libkshark-model.h
@@ -89,7 +89,7 @@ void ksmodel_shift_forward(struct kshark_trace_histo *histo, size_t n);
 
 void ksmodel_shift_backward(struct kshark_trace_histo *histo, size_t n);
 
-void ksmodel_jump_to(struct kshark_trace_histo *histo, size_t ts);
+void ksmodel_jump_to(struct kshark_trace_histo *histo, uint64_t ts);
 
 void ksmodel_zoom_out(struct kshark_trace_histo *histo,
 		      double r, int mark);
